python/test-pdf-text-extractor/requirements.txt
PyPDF2
#PyMuPDF
#fitz
pdfminer.six
pdfplumber
tabula-py
